Added an Array of Reviews
an array of review was created and a function
that set the image, name and position, an indexNumber
was also initialized to keep track of the object in the
array.
The function parameterized, when the function is called
and  indexNumber is passed to it as parameter it display details
of the object.
the function was called on all the buttons by using event listener
to listen to the click event on the botton.